I've added (related with Parking app) - Fixed bugs with getting the ubicaciones (applied to Sanitarios too) - Testing controllers * Pending POST request to parkmovimiento

+++ src/main/java/com/fernando9825/alcaldiasvrestapi/models/services/ParkTarifaImpl.java
@@ -40,7 +40,7 @@ public class ParkTarifaImpl implements IParkTarifaService {
     public List<Parktarifa> getAllParkTarifasByInstitucionIdAndUbicacion(short institucionId, int ubicacion) {
         return this.parkTarifaDao.findAllByInstitucionAndUbicacion(
                 this.institucionDao.findById(institucionId),
-                this.parkUbicacionDao.findById((short) ubicacion)
+                this.parkUbicacionDao.findById(ubicacion)
         );
     }
 }
